# ENH: Add find/rfind ufuncs for unicode and byte dtypes