--- a/numpy/_core/src/umath/string_buffer.h
+++ b/numpy/_core/src/umath/string_buffer.h
@@ -1,0 +1,209 @@
+#ifndef _NPY_CORE_SRC_UMATH_STRING_BUFFER_H_
+#define _NPY_CORE_SRC_UMATH_STRING_BUFFER_H_
+
+#include <Python.h>
+#include <cstddef>
+
+#define NPY_NO_DEPRECATED_API NPY_API_VERSION
+#define _MULTIARRAYMODULE
+#define _UMATHMODULE
+
+#include "numpy/ndarraytypes.h"
+
+#define CHECK_OVERFLOW(index) if (buf + (index) >= after) return 0
+#define MSB(val) ((val) >> 7 & 1)
+
+
+enum class ENCODING {
+    ASCII, UTF32
+};
+
+
+template <ENCODING enc>
+inline npy_ucs4
+getchar(const unsigned char *buf, int *bytes);
+
+
+template <>
+inline npy_ucs4
+getchar<ENCODING::ASCII>(const unsigned char *buf, int *bytes)
+{
+    *bytes = 1;
+    return (npy_ucs4) *buf;
+}
+
+
+template <>
+inline npy_ucs4
+getchar<ENCODING::UTF32>(const unsigned char *buf, int *bytes)
+{
+    *bytes = 4;
+    return *(npy_ucs4 *)buf;
+}
+
+
+template <ENCODING enc>
+struct Buffer {
+    char *buf;
+    char *after;
+
+    inline Buffer<enc>() {
+        buf = after = NULL;
+    }
+
+    inline Buffer<enc>(char *buf_, int elsize_) {
+        buf = buf_;
+        after = buf_ + elsize_;
+    }
+
+    inline npy_int64 length() {
+        Buffer tmp(after, 0);
+        tmp--;
+        while (tmp >= *this && *tmp == '\0') {
+            tmp--;
+        }
+        return (npy_int64) (tmp - *this + 1);
+    }
+
+    inline Buffer<enc>& operator+=(npy_int64 rhs) {
+        switch (enc) {
+        case ENCODING::ASCII:
+            buf += rhs;
+            break;
+        case ENCODING::UTF32:
+            buf += rhs * sizeof(npy_ucs4);
+            break;
+        }
+        return *this;
+    }
+
+    inline Buffer<enc>& operator-=(npy_int64 rhs) {
+        switch (enc) {
+        case ENCODING::ASCII:
+            buf -= rhs;
+            break;
+        case ENCODING::UTF32:
+            buf -= rhs * sizeof(npy_ucs4);
+            break;
+        }
+        return *this;
+    }
+
+    inline Buffer<enc>& operator++() {
+        *this += 1;
+        return *this;
+    }
+
+    inline Buffer<enc> operator++(int) {
+        Buffer<enc> old = *this;
+        operator++();
+        return old; 
+    }
+
+    inline Buffer<enc>& operator--() {
+        *this -= 1;
+        return *this;
+    }
+
+    inline Buffer<enc> operator--(int) {
+        Buffer<enc> old = *this;
+        operator--();
+        return old; 
+    }
+
+    inline npy_ucs4 operator*() {
+        int bytes;
+        return getchar<enc>((unsigned char *) buf, &bytes);
+    }
+
+    inline npy_ucs4 operator[](size_t index) {
+        int bytes;
+        switch (enc) {
+        case ENCODING::ASCII:
+            CHECK_OVERFLOW(index);
+            return getchar<enc>((unsigned char *) (buf + index), &bytes);
+        case ENCODING::UTF32:
+            CHECK_OVERFLOW(index * sizeof(npy_ucs4));
+            return getchar<enc>((unsigned char *) (buf + index * sizeof(npy_ucs4)), &bytes);
+        }
+    }
+
+    inline Buffer<enc> fast_memchr(npy_ucs4 ch, int len) {
+        switch (enc) {
+        case ENCODING::ASCII:
+            buf = (char *) memchr(buf, ch, len);
+            return *this;
+        case ENCODING::UTF32:
+            buf = (char *) wmemchr((wchar_t *) buf, ch, len);
+            return *this;
+        }
+    }
+
+    inline int fast_memcmp(Buffer<enc> other, int len) {
+        return memcmp(buf, other.buf, len);
+    }
+};
+
+
+template <ENCODING enc>
+inline Buffer<enc> operator+(Buffer<enc> lhs, npy_int64 rhs) {
+    lhs += rhs;
+    return lhs;
+}
+
+
+template <ENCODING enc>
+inline std::ptrdiff_t operator-(Buffer<enc> lhs, Buffer<enc> rhs) {
+    switch (enc) {
+    case ENCODING::ASCII:
+        return lhs.buf - rhs.buf;
+    case ENCODING::UTF32:
+        return (lhs.buf - rhs.buf) / (std::ptrdiff_t) sizeof(npy_ucs4);
+    }
+}
+
+
+template <ENCODING enc>
+inline Buffer<enc> operator-(Buffer<enc> lhs, npy_int64 rhs) {
+    lhs -= rhs;
+    return lhs;
+}
+
+
+template <ENCODING enc>
+inline bool operator==(Buffer<enc> lhs, Buffer<enc> rhs) {
+    return lhs.buf == rhs.buf;
+}
+
+
+template <ENCODING enc>
+inline bool operator!=(Buffer<enc> lhs, Buffer<enc> rhs) {
+    return !(rhs == lhs);
+}
+
+
+template <ENCODING enc>
+inline bool operator<(Buffer<enc> lhs, Buffer<enc> rhs) {
+    return lhs.buf < rhs.buf;
+}
+
+
+template <ENCODING enc>
+inline bool operator>(Buffer<enc> lhs, Buffer<enc> rhs) {
+    return rhs < lhs;
+}
+
+
+template <ENCODING enc>
+inline bool operator<=(Buffer<enc> lhs, Buffer<enc> rhs) {
+    return !(lhs > rhs);
+}
+
+
+template <ENCODING enc>
+inline bool operator>=(Buffer<enc> lhs, Buffer<enc> rhs) {
+    return !(lhs < rhs);
+}
+
+
+#endif /* _NPY_CORE_SRC_UMATH_STRING_BUFFER_H_ */