--- a/numpy/_core/src/umath/string_ufuncs.cpp
+++ b/numpy/_core/src/umath/string_ufuncs.cpp
@@ -16,6 +16,8 @@
 #include "convert_datatype.h"
 
 #include "string_ufuncs.h"
+#include "string_fastsearch.h"
+#include "string_buffer.h"
 
 
 template <typename character>
@@ -64,6 +66,51 @@ get_length(const character *str, int elsize)
     return d - str + 1;
 }
 
+
+template <ENCODING enc>
+static inline Py_ssize_t
+findslice(Buffer<enc> str, Py_ssize_t str_len,
+          Buffer<enc> sub, Py_ssize_t sub_len,
+          Py_ssize_t offset)
+{
+    Py_ssize_t pos;
+
+    if (sub_len == 0) {
+        return offset;
+    }
+
+    pos = fastsearch<enc>(str, str_len, sub, sub_len, -1, FAST_SEARCH);
+
+    if (pos >= 0) {
+        pos += offset;
+    }
+
+    return pos;
+}
+
+
+template <ENCODING enc>
+static inline Py_ssize_t
+rfindslice(Buffer<enc> str, Py_ssize_t str_len,
+           Buffer<enc> sub, Py_ssize_t sub_len,
+           Py_ssize_t offset)
+{
+    Py_ssize_t pos;
+
+    if (sub_len == 0) {
+        // Note: This won't work for utf-8 and should be fixed before adding
+        // support for stringdtype.
+        return str_len + offset;
+    }
+
+    pos = fastsearch<enc>(str, str_len, sub, sub_len, -1, FAST_RSEARCH);
+
+    if (pos >= 0) {
+        pos += offset;
+    }
+
+    return pos;
+}
 
 /*
  * Compare two strings of different length.  Note that either string may be
@@ -157,6 +204,93 @@ string_isalpha(const character *str, int elsize)
         str++;
     }
     return (npy_bool) 1;
+}
+
+/* helper to fixup start/end slice values */
+static inline void
+adjust_offsets(npy_int64 *start, npy_int64 *end, npy_int64 len)
+{
+    if (*end > len) {
+        *end = len;
+    } else if (*end < 0) {
+        *end += len;
+        if (*end < 0) {
+            *end = 0;
+        }
+    }
+
+    if (*start < 0) {
+        *start += len;
+        if (*start < 0) {
+            *start = 0;
+        }
+    }
+}
+
+template <ENCODING enc>
+static inline npy_long
+string_find(char *str1, int elsize1, char *str2, int elsize2,
+            npy_int64 start, npy_int64 end)
+{
+    Buffer<enc> buf1(str1, elsize1);
+    Buffer<enc> buf2(str2, elsize2);
+
+    npy_int64 len1 = buf1.length();
+    npy_int64 len2 = buf2.length();
+
+    adjust_offsets(&start, &end, len1);
+    if (end - start < len2) {
+        return (npy_long) -1;
+    }
+
+    if (len2 == 1) {
+        npy_ucs4 ch = *buf2;
+        npy_long result = (npy_long) findchar<enc>(buf1 + start, end - start, ch);
+        if (result == -1) {
+            return (npy_long) -1;
+        } else {
+            return result + (npy_long) start;
+        }
+    }
+
+    return (npy_long) findslice<enc>(buf1 + start,
+                                     end - start,
+                                     buf2,
+                                     len2,
+                                     start);
+}
+
+template <ENCODING enc>
+static inline npy_long
+string_rfind(char *str1, int elsize1, char *str2, int elsize2,
+             npy_int64 start, npy_int64 end)
+{
+    Buffer<enc> buf1(str1, elsize1);
+    Buffer<enc> buf2(str2, elsize2);
+
+    npy_int64 len1 = buf1.length();
+    npy_int64 len2 = buf2.length();
+
+    adjust_offsets(&start, &end, len1);
+    if (end - start < len2) {
+        return (npy_long) -1;
+    }
+
+    if (len2 == 1) {
+        npy_ucs4 ch = *buf2;
+        npy_long result = (npy_long) rfindchar(buf1 + start, end - start, ch);
+        if (result == -1) {
+            return (npy_long) -1;
+        } else {
+            return result + (npy_long) start;
+        }
+    }
+
+    return (npy_long) rfindslice(buf1 + start,
+                                 end - start,
+                                 buf2,
+                                 len2,
+                                 start);
 }
 
 
@@ -296,9 +430,138 @@ string_isalpha_loop(PyArrayMethod_Context *context,
 }
 
 
+template<ENCODING enc>
+static int
+string_find_loop(PyArrayMethod_Context *context,
+        char *const data[], npy_intp const dimensions[],
+        npy_intp const strides[], NpyAuxData *NPY_UNUSED(auxdata))
+{
+    int elsize1 = context->descriptors[0]->elsize;
+    int elsize2 = context->descriptors[1]->elsize;
+
+    char *in1 = data[0];
+    char *in2 = data[1];
+    char *in3 = data[2];
+    char *in4 = data[3];
+    char *out = data[4];
+
+    npy_intp N = dimensions[0];
+
+    while (N--) {
+        npy_long idx = string_find<enc>(in1, elsize1, in2, elsize2,
+                                         *(npy_int64 *)in3, *(npy_int64 *)in4);
+        *(npy_long *)out = idx;
+
+        in1 += strides[0];
+        in2 += strides[1];
+        in3 += strides[2];
+        in4 += strides[3];
+        out += strides[4];
+    }
+    return 0;
+}
+
+
+template<ENCODING enc>
+static int
+string_rfind_loop(PyArrayMethod_Context *context,
+        char *const data[], npy_intp const dimensions[],
+        npy_intp const strides[], NpyAuxData *NPY_UNUSED(auxdata))
+{
+    int elsize1 = context->descriptors[0]->elsize;
+    int elsize2 = context->descriptors[1]->elsize;
+
+    char *in1 = data[0];
+    char *in2 = data[1];
+    char *in3 = data[2];
+    char *in4 = data[3];
+    char *out = data[4];
+
+    npy_intp N = dimensions[0];
+
+    while (N--) {
+        npy_long idx = string_rfind<enc>(in1, elsize1, in2, elsize2,
+                                          *(npy_int64 *)in3, *(npy_int64 *)in4);
+        *(npy_long *)out = idx;
+
+        in1 += strides[0];
+        in2 += strides[1];
+        in3 += strides[2];
+        in4 += strides[3];
+        out += strides[4];
+    }
+    return 0;
+}
+
+
+/* Resolve descriptors & promoter functions */
+
+static NPY_CASTING
+string_addition_resolve_descriptors(
+        PyArrayMethodObject *NPY_UNUSED(self),
+        PyArray_DTypeMeta *NPY_UNUSED(dtypes[3]),
+        PyArray_Descr *given_descrs[3],
+        PyArray_Descr *loop_descrs[3],
+        npy_intp *NPY_UNUSED(view_offset))
+{
+    loop_descrs[0] = NPY_DT_CALL_ensure_canonical(given_descrs[0]);
+    if (loop_descrs[0] == NULL) {
+        return _NPY_ERROR_OCCURRED_IN_CAST;
+    }
+
+    loop_descrs[1] = NPY_DT_CALL_ensure_canonical(given_descrs[1]);
+    if (loop_descrs[1] == NULL) {
+        return _NPY_ERROR_OCCURRED_IN_CAST;
+    }
+
+    loop_descrs[2] = PyArray_DescrNew(loop_descrs[0]);
+    if (loop_descrs[2] == NULL) {
+        return _NPY_ERROR_OCCURRED_IN_CAST;
+    }
+    loop_descrs[2]->elsize += loop_descrs[1]->elsize;
+
+    return NPY_NO_CASTING;
+}
+
+static int
+string_find_rfind_promoter(PyUFuncObject *NPY_UNUSED(ufunc),
+        PyArray_DTypeMeta *op_dtypes[], PyArray_DTypeMeta *signature[],
+        PyArray_DTypeMeta *new_op_dtypes[])
+{
+    Py_INCREF(op_dtypes[0]);
+    new_op_dtypes[0] = op_dtypes[0];
+    Py_INCREF(op_dtypes[1]);
+    new_op_dtypes[1] = op_dtypes[1];
+    new_op_dtypes[2] = PyArray_DTypeFromTypeNum(NPY_INT64);
+    new_op_dtypes[3] = PyArray_DTypeFromTypeNum(NPY_INT64);
+    new_op_dtypes[4] = PyArray_DTypeFromTypeNum(NPY_LONG);
+    return 0;
+}
+
+
 /*
  * Machinery to add the string loops to the existing ufuncs.
  */
+
+static int
+install_promoter(PyObject *umath, const char *ufunc_name,
+                 PyObject *promoterinfo)
+{
+    PyObject *name = PyUnicode_FromString(ufunc_name);
+    if (name == nullptr) {
+        return -1;
+    }
+    PyObject *ufunc = PyObject_GetItem(umath, name);
+    Py_DECREF(name);
+    if (ufunc == nullptr) {
+        return -1;
+    }
+
+    int res = PyUFunc_AddLoop((PyUFuncObject *)ufunc, promoterinfo, 0);
+    Py_DECREF(ufunc);
+    return res;
+}
+
 
 /*
  * This function replaces the strided loop with the passed in one,
@@ -319,7 +582,7 @@ add_loop(PyObject *umath, const char *ufunc_name,
     }
     spec->slots[0].pfunc = (void *)loop;
 
-    int res = PyUFunc_AddLoopFromSpec(ufunc, spec);
+    int res = PyUFunc_AddLoopFromSpec_int(ufunc, spec, 1);
     Py_DECREF(ufunc);
     return res;
 }
@@ -401,127 +664,92 @@ init_comparison(PyObject *umath)
 }
 
 
-static NPY_CASTING
-string_addition_resolve_descriptors(
-        PyArrayMethodObject *NPY_UNUSED(self),
-        PyArray_DTypeMeta *NPY_UNUSED(dtypes[3]),
-        PyArray_Descr *given_descrs[3],
-        PyArray_Descr *loop_descrs[3],
-        npy_intp *NPY_UNUSED(view_offset))
-{
-    loop_descrs[0] = NPY_DT_CALL_ensure_canonical(given_descrs[0]);
-    if (loop_descrs[0] == NULL) {
-        return _NPY_ERROR_OCCURRED_IN_CAST;
-    }
-
-    loop_descrs[1] = NPY_DT_CALL_ensure_canonical(given_descrs[1]);
-    if (loop_descrs[1] == NULL) {
-        return _NPY_ERROR_OCCURRED_IN_CAST;
-    }
-
-    loop_descrs[2] = PyArray_DescrNew(loop_descrs[0]);
-    if (loop_descrs[2] == NULL) {
-        return _NPY_ERROR_OCCURRED_IN_CAST;
-    }
-    loop_descrs[2]->elsize += loop_descrs[1]->elsize;
-
-    return NPY_NO_CASTING;
-}
-
-
 static int
-init_add(PyObject *umath)
+init_ufunc(PyObject *umath, const char *name, const char *specname, int nin, int nout,
+           NPY_TYPES *typenums, PyArrayMethod_StridedLoop *stringmethod,
+           PyArrayMethod_StridedLoop unicodemethod,
+           resolve_descriptors_function resolve_descriptors,
+           promoter_function promoter)
 {
     int res = -1;
 
-    PyArray_DTypeMeta *String = PyArray_DTypeFromTypeNum(NPY_STRING);
-    PyArray_DTypeMeta *Unicode = PyArray_DTypeFromTypeNum(NPY_UNICODE);
+    PyArray_DTypeMeta **dtypes = (PyArray_DTypeMeta **) PyMem_Malloc(
+        (nin + nout) * sizeof(PyArray_DTypeMeta *));
+    if (dtypes == NULL) {
+        PyErr_NoMemory();
+        return -1;
+    }
 
-    /* We start with the string loops: */
-    PyArray_DTypeMeta *dtypes[] = {String, String, String};
-    /*
-     * We only have one loop right now, the strided one.  The default type
-     * resolver ensures native byte order/canonical representation.
-     */
-    PyType_Slot slots[] = {
-        {NPY_METH_strided_loop, nullptr},
-        {NPY_METH_resolve_descriptors, (void *) &string_addition_resolve_descriptors},
-        {0, nullptr}
-    };
+    for (int i = 0; i < nin+nout; i++) {
+        dtypes[i] = PyArray_DTypeFromTypeNum(typenums[i]);
+    }
+
+    PyType_Slot slots[3];
+    slots[0] = {NPY_METH_strided_loop, nullptr};
+    slots[2] = {0, nullptr};
+    if (resolve_descriptors != NULL) {
+        slots[1] = {NPY_METH_resolve_descriptors, (void *) resolve_descriptors};
+    } else {
+        slots[1] = {0, nullptr};
+    }
 
     PyArrayMethod_Spec spec = {};
-    spec.name = "templated_add";
-    spec.nin = 2;
-    spec.nout = 1;
+    spec.name = specname;
+    spec.nin = nin;
+    spec.nout = nout;
     spec.dtypes = dtypes;
     spec.slots = slots;
     spec.flags = NPY_METH_NO_FLOATINGPOINT_ERRORS;
 
     /* All String loops */
-    if (add_loop(umath, "add", &spec, string_add_loop<npy_byte>) < 0) {
+    if (add_loop(umath, name, &spec, stringmethod) < 0) {
         goto finish;
     }
 
     /* All Unicode loops */
-    dtypes[0] = Unicode;
-    dtypes[1] = Unicode;
-    dtypes[2] = Unicode;
-    if (add_loop(umath, "add", &spec, string_add_loop<npy_ucs4>) < 0) {
+    for (int i = 0; i < nin+nout; i++) {
+        if (typenums[i] == NPY_STRING) {
+            Py_DECREF(dtypes[i]);
+            dtypes[i] = PyArray_DTypeFromTypeNum(NPY_UNICODE);
+        }
+    }
+    if (add_loop(umath, name, &spec, unicodemethod) < 0) {
         goto finish;
+    }
+
+    if (promoter != NULL) {
+        PyObject *promoter_obj = PyCapsule_New((void *) promoter, "numpy._ufunc_promoter", NULL);
+        if (promoter_obj == NULL) {
+            goto finish;
+        }
+
+        PyObject *dtypes_tuple = PyTuple_New(nin + nout);
+        if (dtypes_tuple == NULL) {
+            Py_DECREF(promoter_obj);
+            goto finish;
+        }
+        for (int i = 0; i < nin + nout; i++) {
+            PyTuple_SET_ITEM(dtypes_tuple, i, Py_None);
+        }
+
+        PyObject *info = PyTuple_Pack(2, dtypes_tuple, promoter_obj);
+        Py_DECREF(dtypes_tuple);
+        Py_DECREF(promoter_obj);
+        if (info == NULL) {
+            goto finish;
+        }
+
+        if (install_promoter(umath, name, info) < 0) {
+            goto finish;
+        }
     }
 
     res = 0;
   finish:
-    Py_DECREF(String);
-    Py_DECREF(Unicode);
-    return res;
-}
-
-
-static int
-init_isalpha(PyObject *umath)
-{
-    int res = -1;
-    /* NOTE: This should receive global symbols? */
-    PyArray_DTypeMeta *String = PyArray_DTypeFromTypeNum(NPY_STRING);
-    PyArray_DTypeMeta *Unicode = PyArray_DTypeFromTypeNum(NPY_UNICODE);
-    PyArray_DTypeMeta *Bool = PyArray_DTypeFromTypeNum(NPY_BOOL);
-
-    /* We start with the string loops: */
-    PyArray_DTypeMeta *dtypes[] = {String, Bool};
-    /*
-     * We only have one loop right now, the strided one.  The default type
-     * resolver ensures native byte order/canonical representation.
-     */
-    PyType_Slot slots[] = {
-        {NPY_METH_strided_loop, nullptr},
-        {0, nullptr}
-    };
-
-    PyArrayMethod_Spec spec = {};
-    spec.name = "templated_string_isalpha";
-    spec.nin = 1;
-    spec.nout = 1;
-    spec.dtypes = dtypes;
-    spec.slots = slots;
-    spec.flags = NPY_METH_NO_FLOATINGPOINT_ERRORS;
-
-    /* All String loops */
-    if (add_loop(umath, "isalpha", &spec, string_isalpha_loop<npy_byte>) < 0) {
-        goto finish;
+    for (int i = 0; i < nin+nout; i++) {
+        Py_DECREF(dtypes[i]);
     }
-
-    /* All Unicode loops */
-    dtypes[0] = Unicode;
-    if (add_loop(umath, "isalpha", &spec, string_isalpha_loop<npy_ucs4>) < 0) {
-        goto finish;
-    }
-
-    res = 0;
-  finish:
-    Py_DECREF(String);
-    Py_DECREF(Unicode);
-    Py_DECREF(Bool);
+    PyMem_Free((void *) dtypes);
     return res;
 }
 
@@ -529,15 +757,41 @@ init_isalpha(PyObject *umath)
 NPY_NO_EXPORT int
 init_string_ufuncs(PyObject *umath)
 {
+    NPY_TYPES dtypes[] = {NPY_STRING, NPY_STRING, NPY_STRING, NPY_STRING, NPY_STRING};
+
     if (init_comparison(umath) < 0) {
         return -1;
     }
 
-    if (init_add(umath) < 0) {
+    if (init_ufunc(umath, "add", "templated_string_add", 2, 1, dtypes,
+                   string_add_loop<npy_byte>, string_add_loop<npy_ucs4>,
+                   string_addition_resolve_descriptors, NULL) < 0) {
         return -1;
     }
 
-    if (init_isalpha(umath) < 0) {
+    dtypes[0] = NPY_STRING;
+    dtypes[1] = NPY_BOOL;
+    if (init_ufunc(umath, "isalpha", "templated_string_isalpha", 1, 1, dtypes,
+                   string_isalpha_loop<npy_byte>, string_isalpha_loop<npy_ucs4>,
+                   NULL, NULL) < 0) {
+        return -1;
+    }
+
+    dtypes[0] = dtypes[1] = NPY_STRING;
+    dtypes[2] = dtypes[3] = NPY_INT64;
+    dtypes[4] = NPY_LONG;
+    if (init_ufunc(umath, "find", "templated_string_find", 4, 1, dtypes,
+                   string_find_loop<ENCODING::ASCII>, string_find_loop<ENCODING::UTF32>,
+                   NULL, string_find_rfind_promoter) < 0) {
+        return -1;
+    }
+
+    dtypes[0] = dtypes[1] = NPY_STRING;
+    dtypes[2] = dtypes[3] = NPY_INT64;
+    dtypes[4] = NPY_LONG;
+    if (init_ufunc(umath, "rfind", "templated_string_rfind", 4, 1, dtypes,
+                   string_rfind_loop<ENCODING::ASCII>, string_rfind_loop<ENCODING::UTF32>,
+                   NULL, string_find_rfind_promoter) < 0) {
         return -1;
     }