… benchmarks/benchmarks/bench_core.py
… doc/release/upcoming_changes/24868.new_feature.rst
… numpy/_core/src/umath/string_fastsearch.h
… numpy/_core/tests/test_ufunc.py