@lysnikolaou 2023-10-18
Benchmark results:

```
| Change   | Before [19bfa3ff] <main>   | After [b35de314] <string-ufuncs-find-rfind>   |   Ratio | Benchmark (Parameter)                                |
|----------|----------------------------|-----------------------------------------------|---------|------------------------------------------------------|
| -        | 17.1±0.08μs                | 9.24±0.02μs                                   |    0.54 | bench_core.NumPyChar.time_find_small_list_big_string |
| -        | 5.67±0.03ms                | 15.5±0.05μs                                   |    0    | bench_core.NumPyChar.time_find_big_list_small_string |
```