@lysnikolaou 2023-10-18
>  Do we you really need to vendor that file with large changes? Any chance of vendoring it without changes?

Unfortunately, for this to play along with `string_ufuncs.cpp` and the templating mechanism it uses, the changes are needed. This file has been there part of the CPython source tree for a long time with very minimal changes, in case that makes you less concerned.