@lysnikolaou 2023-10-19
How should I go about fixing the test failure?

```
numpy._core._exceptions._UFuncNoLoopError:
    ufunc 'find' did not contain a loop with signature matching types
    (<class 'numpy.dtypes.StrDType'>,
     <class 'numpy.dtypes.StrDType'>,
     <class 'numpy.dtypes.Int64DType'>,
     <class 'numpy.dtypes.Int64DType'>) -> <class 'numpy.dtypes.Int32DType'>
```

Add another loop? Downcast the result? While I think I've understood a lot about how to handle input arguments dtypes and type resolving, output dtypes still confuse me a bit for some reason.