@ngoldbaum 2023-10-24
I left a comment on the last commit, should probably have been here: https://github.com/numpy/numpy/commit/fcd600ac0944e4e1925216e1e5c7c0fcbed8bd1f#r130812759