@lysnikolaou 2023-10-27
@seberg @ngoldbaum I've reworked this PR to use the buffer approach suggested by @ngoldbaum. Unfortunately, the tests all fail with the same kind of error:

```
[gw1] linux -- Python 3.11.6 /opt/hostedtoolcache/Python/3.11.6/x64/bin/python

    def test_find_access_past_buffer():
        # This checks that no read past the string buffer occurs in
        # string_fastsearch.h. The READC macro is there to check this.
        # To see it in action, you can redefine READC to just read the
        # i'th character of the buffer and this test will produce an
        # 'Invalid read' if run under valgrind.
        arr = np.array([b'abcd', b'ebcd'])
>       result = np._core.umath.find(arr, b'cde', 0, np.iinfo(np.int64).max)
E       numpy._core._exceptions._UFuncNoLoopError: ufunc 'find' did not contain a loop with signature matching types (<class 'numpy.dtypes.BytesDType'>, <class 'numpy.dtypes.BytesDType'>, <class 'numpy._IntegerAbstractDType'>, <class 'numpy._IntegerAbstractDType'>) -> None
```

This seems completely unrelated to the change I did in the last commits and I can't reproduce it locally either on a Ubuntu machine or an M1 macbook. Any ideas what might be causing it?