@ngoldbaum 2023-10-30
Are the benchmark results still similar to https://github.com/numpy/numpy/pull/24868#issuecomment-1768066613?