@lysnikolaou 2023-10-30
Yeah, sorry, I forgot to post the updated benchmark results.

```
| Change   | Before [cdfbdf42] <main>   | After [03d737dc] <string-ufuncs-find-rfind>   |   Ratio | Benchmark (Parameter)                                |
|----------|----------------------------|-----------------------------------------------|---------|------------------------------------------------------|
| -        | 17.1±0.2μs                 | 9.45±0.3μs                                    |    0.55 | bench_core.NumPyChar.time_find_small_list_big_string |
| -        | 5.73±0.2ms                 | 17.2±0.4μs                                    |    0    | bench_core.NumPyChar.time_find_big_list_small_string |
```