@ngoldbaum 2023-10-30
> Maybe we can just remove the utf-8 stuff? Or put a big comment somewhere that the utf-8 stuff was intended to be used in the near future, but is not usable yet.


Removing it is fine with me.

I'm happy to benchmark whether implicit conversions to UCS-4 arrays is faster than just plugging in the Buffer class when I do that.