@lysnikolaou 2023-10-30
Removed the UTF-8 stuff and addressed all review comments.