@seberg 2023-10-31
> I'm happy to benchmark whether implicit conversions to UCS-4 arrays is faster than just plugging in the Buffer class when I do that.

Yeah, I am not too worried about it.  My gut feeling is that the simplest solution for many algorithms will be to use the bytes directly (for whitespace/isalpha, etc. of course not, but those are clean single pass algorithms anyway).
For find/rfind you will have to do translate the slicing into byte-offets at the start and the result byte-offset back to character offsets when done.

Btw. I don't think it is useful enough at this point, but the code could in principle be optimized a bit for the likely case that needle is a constant: we can re-use the bloom filter from the previous iteration.