@seberg 2023-11-02
@lysnikolaou sorry, I don't want to fret about a few style nits, and I can't see a better path than vendoring the Python stuff (I really don't love how much code it is for something we don't do often, but...).
So I think, given that Nathan is also around to help with maintaining thing, I think we can just put it in.

There is one problem now, unfortunately.  I *suspect* the tests will fail after merging the change to what the default integer is.  My suggestion would be that you squash all current commits into one (or few) and then add a single commit to fix all `NPY_LONG` to `NPY_DEFAULT_INT` (or `intp` if easier), hopefully that will make any test pass?

That way, the PR isn't directly tied to any possible revert, etc. due to the intp changes.